refactor: unify duplicated test helpers into tests/common
src/testutil.rs and tests/common/mod.rs duplicated ~120 lines of
helpers (TestContext, create_overlay_dir, ...) and had already started
drifting. tests/common/mod.rs is now the single source of truth: unit
tests reach it through a #[path]-mapped crate::testutil module in
lib.rs, integration tests keep using mod common.

SourceTestContext moves into tests/cli.rs (its only consumer) because
cargo_bin_cmd! needs CARGO_BIN_EXE_* env vars that cargo only provides
to integration-test builds.

Fixes #367
